Improved watched posts notifications to consider user notifications settings
