+++ src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Email/Posting/ICommentEmailNotificationService.cs
@@ -1,0 +1,9 @@
+﻿using Shrooms.DataTransferObjects.Models.Wall.Posts.Comments;
+
+namespace Shrooms.Domain.Services.Email.Posting
+{
+    public interface ICommentEmailNotificationService
+    {
+        void SendEmailNotification(CommentCreatedDTO commentDto);
+    }
+}
