+++ src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Wall/Posts/Comments/ICommentService.cs
@@ -6,11 +6,15 @@ namespace Shrooms.Domain.Services.Wall.Posts.Comments
     public interface ICommentService
     {
         void EditComment(EditCommentDTO commentDto);
+
         CommentCreatedDTO CreateComment(NewCommentDTO commentDto);
+
         void ToggleLike(int commentId, UserAndOrganizationDTO userOrg);
+
         void DeleteComment(int id, UserAndOrganizationDTO userOrg);
+
         void HideComment(int id, UserAndOrganizationDTO userOrg);
-        void DeleteCommentsByPost(int postId, UserAndOrganizationDTO userAndOrg);
-        bool IsPostAuthorAppNotificationsEnabled(string userId);
+
+        void DeleteCommentsByPost(int postId);
     }
 }
