+++ src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Wall/Posts/IPostService.cs
@@ -14,7 +14,7 @@ namespace Shrooms.Domain.Services.Wall.Posts
         void DeleteWallPost(int postId, UserAndOrganizationDTO userOrg);
         void HideWallPost(int postId, UserAndOrganizationDTO userOrg);
         void ToggleWatch(int postId, UserAndOrganizationDTO userAndOrg, bool shouldWatch);
-        IEnumerable<string> GetPostWatchersIds(int postId);
-        IEnumerable<ApplicationUser> GetPostWatchers(int postId);
+        IEnumerable<string> GetPostWatchersForAppNotifications(int postId);
+        IEnumerable<ApplicationUser> GetPostWatchersForEmailNotifications(int postId);
     }
 }
